makes read_jsonlines a generator and moves to utils
