+++ test/test_utils.py
@@ -3,7 +3,8 @@ import boto3
 from moto import mock_s3
 from botocore.exceptions import ClientError
 
-from croissant.utils import nested_get_item, s3_get_object
+from croissant.utils import (nested_get_item, s3_get_object,
+                             read_jsonlines)
 
 
 @pytest.mark.parametrize(
@@ -71,3 +72,42 @@ def test_s3_fails_not_exist():
     with pytest.raises(ClientError) as e:
         s3_get_object("s3://mybucket/my/nonexistentfile.json")
         assert e.response["Error"]["Code"] == "NoSuchKey"
+
+
+@pytest.mark.parametrize(
+    "body, expected",
+    [
+        (b'{"a": 1, "b": 3}\n{"b": 2}', [{"a": 1, "b": 3}, {"b": 2}]),
+        (b'{"a": 1}', [{"a": 1}]),
+        (b'', []),
+    ]
+)
+def test_read_jsonlines_file(tmp_path, body, expected):
+    with open(tmp_path / "filename", "wb") as f:
+        f.write(body)
+    reader = read_jsonlines(tmp_path / "filename")
+    response = []
+    for record in reader:
+        response.append(record)
+    assert expected == response
+
+
+@mock_s3
+@pytest.mark.parametrize(
+    "body, expected",
+    [
+        (b'{"a": 1, "b": 3}\n{"b": 2}', [{"a": 1, "b": 3}, {"b": 2}]),
+        (b'{"a": 1}', [{"a": 1}]),
+        (b'', []),
+    ]
+)
+def test_read_jsonlines_s3(body, expected):
+    s3 = boto3.client("s3")
+    s3.create_bucket(Bucket="mybucket")
+    s3.put_object(Bucket="mybucket", Key="my/file.json",
+                  Body=body)
+    reader = read_jsonlines("s3://mybucket/my/file.json")
+    response = []
+    for record in reader:
+        response.append(record)
+    assert expected == response
